[FileTypes] Start of split of VideoFrameCache and video effects modifications. The goal is to have separate classes for frame containers to better separate roles, decouple the abstract reader from a specific container, increase testability, and extract thread related code from the Cache container. There will be 3 containers, Cache (for full working zone buffer), PreBuffer and single frame. Ideally the base video reader shouldn't assume which specific container the video reader is using.
- Created the classes and implementation for Cache.
- The specific container member is moved to the reader implementations.

VideoReaderAlwaysCaching: A new abstract reader is created for conveniency to video readers that are always caching all frames.
1. group boiler plate implementation of irrelevant functions.
2. provide standard implementation for cache related operations.
- VideoReaderGif subclass VideoReaderAlwaysCaching.

For video effects, instead of being coupled to the VideoFrameCache, the working zone frames are exposed by the reader as an IWorkingZoneFramesContainer.
This interface declares methods used by video effects like getting the collection of frames, getting a representative image, etc.

In video effects, slightly altered the prototype of ImageProcessor delegate since the processing is in-place.
This allows even more simple code for some filters, as the AForge function can directly be used as the ImageProcessor.