Improving "rendering unscaled" mechanics.
We wait for the first idle event after the video load to start prebuffering and the first working zone update.
This avoids the multiple resize events of the initial screen creation.
(Side effect, the hack to unset image fill mode during screen creation is no longer needed).

During manual resize of the rendering surface, we don't update the decoding size until MouseUp event.
This avoids a potentially costly reset of the prebuffer.

To minimize coupling, the player checks if the image size (or zoom window) is about right to be drawn unscaled.
During manual resizing for example, the decoding size is not updated and we must draw "scaled".

New concept of RenderingZoomWindow, to draw unscaled the zoom window must be relative to the decoding size.

New:
- user can squeeze the rendering surface to go smaller than the original size.
- This allows for better perfs at the expense of the size. The user manages the tradeoff.
- The reference size is made a sticky area for easy sizing.

Refactoring: Zoom window relocation functions simplified.
Bugfix: When trying to load the selection to cache, if the first image fails we were misreporting success.