Drawing tool for coordinate system origin + refactoring.
Reimplementation of the dialog to set the origin of coordinate system in a drawing tool.
The new utility is directly modified on the main image and can be color configured, tracked and has several display modes (cross, axis, grid, graduations ticks).
Access is still done through the main menu.

Refactoring:
- Moved form location code to a static helper.
- Renamed Track to DrawingTrack.
- The Calibration helper (rather than the Line tool) raises an event when calibration is changed.
- Drawings' custom menus is a list of ToolStripItem (instead of ToolStripMenuItem) in order to contain separators.
- Drawing hit during hit test is kept in the HitDrawing property at Metadata level.
This allows for unification of context menus creation and handlers between attached drawings and extra drawings.