Move the camera transformer calls inside Trackability manager.
The points we want to move through camera tracking are the same as the points we move through point tracking.
This way the point coordinates are moved from outside the drawing and all the functions relying on them work transparently (draw, hit test, move).
When we move a point manually we signal it through the same mechanism as for point tracking, but we re-import the value from the drawing (updated by camera tracking in the meantime).