Refactoring Tracking parameters
TrackingParameters is now under Services and is responsible for serialization of the tracking parameters for prefs as well as individual drawings.
Moved All tracking related classes out of the "Measurement" folder into a "Tracking" folder.
Removed TrackingProfile and TrackerParameterUnit classes, replaced by the new TrackingParameters.
Removed the Tracking tab under preferences as it no longer serves any purposes since the default parameters are always updated from last usage.
Removed the duplicate variables in TrackerBlock and use the parameters object directly.
DrawingTrack is no longer an IScalable since the concept of initializing the tracking parameters as a percentage of image size doesn't exist anymore.