WIP tracking panel.
Remove zooming parameter from hit testing and moving of drawings. (This is an attempt at converging the two image-to-viewport transformer classes).
Fix issue with scale of hit test in the tracking mini viewport.
Perfs in tracking mini viewport. Only update main viewport when the manipulation is finished.
Fix bug that was saving the wrong boolean for IsInteractiveTrack.
Hide recording indicator.