Catch exception when using the portable version from within a synchronized OneDrive directory.
Due to an issue in OneDrive the deletion of the temporary directory fails with Access Denied which caused a crash in both cleanup of auto-save data when closing a screen and in the recovery mechanism when starting up.
This prevents the program from starting as it fails during cleanup of the recovery folder during init.