Catch exception when using the portable version from within a synchronized OneDrive directory.
Due to an issue in OneDrive the deletion of the temporary directory fails with Access Denied which caused a crash in both cleanup of auto-save data when closing a screen and in the recovery mechanism when starting up.
This prevents the program from starting as it fails during cleanup of the recovery folder during init.

Cherry picked from master: 7b8c398ae1714b12bc00821150ace4bacc1e6df8
(2025-03-02)