Added more context menus for annotations management.
- Save as default player annotations
- Save as default capture annotations
- Reload default player annotations
- Reload default capture annotations
- Reload linked annotations

Added name of target file to Save annotations and Reload linked annotations.

Fixed the behavior when loading companion file in the player. We were assigning the default kva path to the working file and any further save action was overwriting the default file.
Fixed cancellation of the file picker when closing a screen and clicking yes to save but it has to do a save as. Previously it was closing the screen, now it cancels the whole operation.